# backend/nodejs/src/types/Note.ts
export interface Note {
  id: string;
  title: string;
  content: string;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface UpdateNoteRequest {
  title?: string;
  content?: string;
  tags?: string[];
}